--- Projects/Project2/Misc/regan_draft/__init__.py
@@ -1,4 +1,0 @@
-from .crossvalidation import cross_validation
-from .bootstrap_biasvar import Confidence_Interval, bias_variance_complexity, bootstrap, bias_variance_analysis
-from .regression import lasso_reg, ridge_reg, OLS_solver, FrankeFunction, create_X, Split_and_Scale, MSE, R2, Plot_FrankeFunction, create_xyz_dataset, plot_ols_complexity, Rolling_Mean
-from .run_and_plot import run_plot_compare, compare_lmd_CV, compare_lmd_BS
